requirements.txt
click>=8.1.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.0.0
dependency-injector>=4.41.0
rich>=13.0.0
python-dotenv>=1.0.0
structlog>=23.0.0
aiohttp>=3.8.0
openai>=1.0.0